cleaned up server side to follow mvc better. Note that view is the react front end
