+++ server/routes/authRoutes.js
@@ -1,0 +1,11 @@
+const express = require('express');
+const router = express.Router();
+const authController = require('../controllers/authController'); // Import the controllers
+
+// Signup route
+router.post('/signup', authController.register);
+
+// Login route
+router.post('/login', authController.login);
+
+module.exports = router;
